requirements.txt
pandas>=2.2
numpy>=1.26
requests>=2.31
python-dotenv>=1.0
PyYAML>=6.0
yfinance>=0.2.50
matplotlib>=3.8
